refactor: import paths for sanity types
