+++ src/components/article/ArticleDetails/Sections/MainImage.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image';
 import { useArticleDetailsStore } from '@/components/article/ArticleDetails/context';
 import { Separator } from '@/components/ui/separator';
 import { client } from '@/sanity/lib/client';
-import { IArticle } from '@/sanity/types/test';
+import { IArticle } from '@/sanity/types';
 
 const builder = imageUrlBuilder(client);
