+++ src/components/project/ProjectListParent.tsx
@@ -2,9 +2,9 @@
 
 import FeaturedProject from '@/components/project/FeaturedProject';
 import ProjectGrid from '@/components/project/ProjectGrid';
-import { ProjectPost } from '@/sanity/types';
+import { IArticle } from '@/sanity/types';
 
-const ProjectListParent = ({ projects }: { projects: ProjectPost[] }) => {
+const ProjectListParent = ({ projects }: { projects: IArticle[] }) => {
   if (!projects) {
     return <div className="text-center">No projects available.</div>;
   }
@@ -19,7 +19,7 @@ const ProjectListParent = ({ projects }: { projects: ProjectPost[] }) => {
           title={mainPost.title}
           coverImage={mainPost.mainImage}
           date={mainPost.publishedAt}
-          excerpt={mainPost.excerpt}
+          excerpt={mainPost.excerpt as string}
           author={mainPost.author}
           slug={mainPost.slug}
         />
